Moving test utils class
