+++ tests/ff_module-button.test.js
@@ -1,5 +1,5 @@
 var React = require('react');
-require('./utils').bootstrapBrowser();
+require('./lib/utils').bootstrapBrowser();
 
 var TestUtils = require('react-addons-test-utils'),
     expect = require('chai').expect,
