+++ tests/lib/utils.js
